Pair of Shoes
Yesterday you found some shoes in the back of your closet. Each shoe is described by two values:

1. type indicates if it's a left or a right shoe;
2. size is the size of the shoe.

Your task is to check whether it is possible to pair the shoes you found in such a way that each pair consists of a right and a left shoe of an equal size.