Crossing Sum
Given a rectangular matrix and integers a and b, consider the union of the a^th row and the b^th (both 0-based) column of the matrix (i.e. all cells that belong either to the a^th row or to the b^th column, or to both). Return sum of all elements of that union.